requirements.txt
requests
paramiko
python-dotenv
openai
